refactor: reorganize user page file structure
- 移除 user.tsx 已改為使用 wrappers.tsx
- 重新命名 pages/user/Wrapper 目錄結構
- 更新相關 component 的引用、參考路徑
