+++ components/pages/user/Wrapper/index.tsx
@@ -1,0 +1,12 @@
+import { ReactNode } from "react";
+import { MainWrapper } from "@/styles/wrappers";
+
+interface WrapperProps {
+  children: ReactNode;
+}
+
+const Wrapper = ({ children }: WrapperProps) => {
+  return <MainWrapper>{children}</MainWrapper>;
+};
+
+export default Wrapper;
