--- styles/user.tsx
@@ -1,8 +1,0 @@
-import styled from "styled-components";
-import { Container1344 } from "@/styles/container";
-
-export const UserContainer = styled(Container1344)`
-  padding: 80px 0;
-  column-gap: 72px;
-  display: flex;
-`;
